requirements.txt
appnope==0.1.0
asgiref==3.2.10
backcall==0.2.0
decorator==4.4.2
Django==3.1.2
django-cors-headers==3.6.0
django-crispy-forms==1.9.2
django-shell-ipynb==0.3.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.4.0
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.2
mysqlclient==2.0.1
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
prompt-toolkit==3.0.8
ptyprocess==0.6.0
Pygments==2.7.2
PyJWT==1.7.1
pytz==2020.1
pyzmq==19.0.2
six==1.15.0
sqlparse==0.4.1
tornado==6.0.4
traitlets==4.3.3
wcwidth==0.2.5
